requirements.txt
# Analytics Business Framework
# Core dependencies for the 5-step Data Flywheel

# Web Dashboard
streamlit>=1.28.0
plotly>=5.18.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# BigQuery
google-cloud-bigquery>=3.13.0
google-cloud-bigquery-storage>=2.22.0
db-dtypes>=1.1.1

# Machine Learning (for Prediction module)
scikit-learn>=1.3.0

# Monte Carlo Simulation
scipy>=1.11.0

# Configuration
python-dotenv>=1.0.0
pydantic>=2.5.0
